add authorization handling for PUT request

+++ src/pages/api/venues/[id].js
@@ -1,3 +1,4 @@
+import { getServerSession } from "next-auth/next";
 import dbConnect from "../../../db/connect";
 import Review from "../../../db/models/Review";
 import { authOptions } from "../auth/[...nextauth]";
@@ -17,6 +18,7 @@ export default async function handler(request, response) {
     return response.status(200).json(reviews);
   }
   if (request.method === "POST") {
+    if (!session) return response.status(401).json({ status: "Unauthorized" });
     try {
       const reviewData = request.body;
       const review = new Review(reviewData);
@@ -28,12 +30,18 @@ export default async function handler(request, response) {
     }
   }
   if (request.method === "PUT") {
+    if (!session.user.id === request.body.user_id) {
+      return response.status(401).json({ status: "Unauthorized" });
+    }
     const reviewToUpdate = await Review.findByIdAndUpdate(id, {
       $set: request.body,
     });
     response.status(200).json(reviewToUpdate);
   }
   if (request.method === "DELETE") {
+    if (!session.user.id === request.body.user_id) {
+      return response.status(401).json({ status: "Unauthorized" });
+    }
     const reviewToDelete = await Review.findByIdAndDelete(id);
     response.status(200).json(reviewToDelete);
   }
